2 smaller changes
- Changed the empty return state PredictionContext to max int (instead fixing it on int 16), like it's done in the Java runtime.
- Converted 2 static init lambdas in the Cpp.stg to normal code, as VS doesn't allow us to access private class members in such lambdas.

+++ runtime/Cpp/runtime/atn/PredictionContext.h
@@ -56,7 +56,7 @@ namespace atn {
     /// Represents $ in an array in full context mode, when $
     /// doesn't mean wildcard: $ + x = [$,x]. Here,
     /// $ = EMPTY_RETURN_STATE.
-    static const int EMPTY_RETURN_STATE = INT16_MAX;
+    static const int EMPTY_RETURN_STATE = INT_MAX;
 
   private:
     static const int INITIAL_HASH = 1;
